requirements.txt
# Core
pyyaml>=6.0
# API
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.0
# Database
sqlalchemy>=2.0
# Testing
pytest>=7.0
pytest-asyncio>=0.21.0
httpx>=0.25.0
# Utilities
python-dateutil>=2.8
requests>=2.31.0
